Moved parse_nested_xml_text_nodes test to y_doc module
